Rename error_messages.py to minor_error_code.py
